added follow and unfollow commands
